Add production UI build pipeline
- Build and package a self-contained `ui/dist` bundle for Docker
- Add production build tests and switch deploy image to the new output
