+++ ui/index.html
@@ -48,5 +48,5 @@
 	<main>Loading...</main>
 
 	<!-- a workaround for bug: https://bugzilla.mozilla.org/show_bug.cgi?id=1803984 -->
-	<script async type = 'module'>import './js/index.js'</script>
+	<script async type = 'module'>import './js/index.dev.js'</script>
 </body>
