+++ ui/package.json
@@ -32,6 +32,7 @@
 		"shared": "cd .. && bun run shared:build && bun run ui:shared",
 		"build:workers": "bun ./build/workers.mts",
 		"build": "bun run shared && tsc && bun run build:workers",
+		"build:prod": "bun ./build/production.mts",
 		"build:tests": "bun ./build/tests.mts",
 		"watch": "bun run shared && tsc --watch",
 		"serve": "bun run shared && bun x tsc --project tsconfig.json && bun ./dev-server.ts",
