+++ ui/ts/index.dev.ts
@@ -1,0 +1,2 @@
+import './liveReload.js'
+import './boot.js'
